+++ packages/banner/src/BannerDismissButton/BannerDismissButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { cx } from '@leafygreen-ui/emotion';
 import XIcon from '@leafygreen-ui/icon/dist/X';
-import IconButton from '@leafygreen-ui/icon-button';
+import { IconButton } from '@leafygreen-ui/icon-button';
 
 import { baseStyles, variantStyles } from './BannerDismissButton.styles';
 import { BannerDismissButtonProps } from './BannerDismissButton.types';