+++ packages/code-editor/src/Panel/Panel.styles.ts
@@ -1,4 +1,4 @@
-import { css } from '@leafygreen-ui/emotion';
+import { css, cx } from '@leafygreen-ui/emotion';
 import { Theme } from '@leafygreen-ui/lib';
 import {
   borderRadius,
@@ -8,29 +8,57 @@ import {
   Variant,
 } from '@leafygreen-ui/tokens';
 
-const PANEL_HEIGHT = 36;
+/**
+ * The Loading div is absolutely positioned inside the code editor. When the panel isn't rendered,
+ * it renders at a `top: 0` position. When the panel is rendered, it needs to render at a
+ * `top: ${PANEL_HEIGHT}px` position so it doesn't overlap with the panel. So this is exported.
+ */
+export const PANEL_HEIGHT = 36;
+
 const MODAL_HEIGHT = 354;
 
-export const getPanelStyles = (theme: Theme) => {
-  return css`
-    background-color: ${color[theme].background[Variant.Secondary][
-      InteractionState.Default
-    ]};
-    height: ${PANEL_HEIGHT}px;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 ${spacing[400]}px 0 ${spacing[300]}px;
-    border: 1px solid
-      ${color[theme].border[Variant.Secondary][InteractionState.Default]};
-    border-bottom: none;
-    border-top-left-radius: ${borderRadius[300]}px;
-    border-top-right-radius: ${borderRadius[300]}px;
-    display: grid;
-    grid-template-columns: auto 1fr auto;
-    grid-template-areas: 'title inner-content buttons';
-  `;
+const getBasePanelStyles = (theme: Theme) => css`
+  background-color: ${color[theme].background[Variant.Secondary][
+    InteractionState.Default
+  ]};
+  height: ${PANEL_HEIGHT}px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 ${spacing[400]}px 0 ${spacing[300]}px;
+  border: 1px solid
+    ${color[theme].border[Variant.Secondary][InteractionState.Default]};
+  border-bottom: none;
+  border-top-left-radius: ${borderRadius[300]}px;
+  border-top-right-radius: ${borderRadius[300]}px;
+  display: grid;
+  grid-template-columns: auto 1fr auto;
+  grid-template-areas: 'title inner-content buttons';
+  width: 100%;
+`;
+
+export const getPanelStyles = ({
+  theme,
+  width,
+  minWidth,
+  maxWidth,
+}: {
+  theme: Theme;
+  width?: string;
+  minWidth?: string;
+  maxWidth?: string;
+}) => {
+  return cx(getBasePanelStyles(theme), {
+    [css`
+      width: ${width};
+    `]: !!width,
+    [css`
+      min-width: ${minWidth};
+    `]: !!minWidth,
+    [css`
+      max-width: ${maxWidth};
+    `]: !!maxWidth,
+  });
 };
 
 export const getPanelTitleStyles = (theme: Theme, baseFontSize: number) => {
@@ -50,6 +78,8 @@ export const getPanelInnerContentStyles = () => {
 export const getPanelButtonsStyles = () => {
   return css`
     grid-area: buttons;
+    display: flex;
+    justify-content: center;
   `;
 };