+++ packages/feature-walls/src/FeatureOverview/FeatureOverview.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef, useState } from 'react';
 
-import Card from '@leafygreen-ui/card';
+import { Card } from '@leafygreen-ui/card';
 import LeafyGreenProvider, {
   useDarkMode,
 } from '@leafygreen-ui/leafygreen-provider';