+++ chat/lg-markdown/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lg-chat/lg-markdown
 
+## 4.2.2
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/typography@22.2.0
+  - @leafygreen-ui/code@20.2.1
+
 ## 4.2.1
 
 ### Patch Changes