+++ packages/code/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leafygreen-ui/code
 
+## 20.2.1
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [c6b4d3f]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/select@16.3.0
+  - @leafygreen-ui/icon@14.6.1
+  - @leafygreen-ui/typography@22.2.0
+  - @leafygreen-ui/button@25.1.2
+  - @leafygreen-ui/hooks@9.2.2
+  - @leafygreen-ui/icon-button@17.1.2
+  - @leafygreen-ui/skeleton-loader@3.0.8
+  - @leafygreen-ui/tooltip@14.2.2
+
 ## 20.2.0
 
 ### Minor Changes