+++ packages/skeleton-loader/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @leafygreen-ui/skeleton-loader
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [c6b4d3f]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/card@13.2.0
+  - @leafygreen-ui/icon@14.6.1
+  - @leafygreen-ui/typography@22.2.0
+
 ## 3.0.7
 
 ### Patch Changes